Quiet the GCC build: fix -Wsign-compare sites, drop GCC -Warray-bounds, silence Boost.Bind pragma
Building with GCC on MSYS2 (and on Linux) produced three kinds of noise
that clang's -Wall does not show:

- 12 -Wsign-compare warnings, all int-vs-size_t. Nine are the thread-pool
  back-pressure loops comparing getQueueSize() with thread_num * 2 (or + 2);
  cast the int side to std::size_t and include <cstddef> where it was
  missing. Env::removeLeftPeaks/removeRightPeaks now compute an int bound
  once (the old size() - num also wrapped when num exceeded the size), and
  EnvSet compares peak_num with an int cast of the theoretical size.

- GCC 12+ "array subscript ... is partly outside array bounds" traces from
  libstdc++'s inlined make_shared/shared_ptr code at -O3: known false
  positives, so add -Wno-array-bounds under GCC only.

- The "#pragma message" from <boost/bind.hpp> in pwiz's
  SpectrumList_mzML.cpp is not a warning and survives -w; define
  BOOST_BIND_GLOBAL_PLACEHOLDERS privately for the pwiz target.

A full GCC 13 Linux build now compiles with no warnings.

+++ src/search/ptmsearch/ptm_search_processor.cpp
@@ -17,6 +17,7 @@
 
 #include <algorithm>
 #include <cmath>
+#include <cstddef>
 
 #include "common/thread/simple_thread_pool.hpp"
 #include "common/util/file_util.hpp"
@@ -163,7 +164,7 @@ void PtmSearchProcessor::process() {
         }
         if (selected_prsm_ptrs.size() > 0) {
           for (size_t i = 0; i < spec_set_ptr_vec.size(); i++) {
-            while (pool_ptr->getQueueSize() >= mng_ptr_->thread_num_ * 2) {
+            while (pool_ptr->getQueueSize() >= static_cast<std::size_t>(mng_ptr_->thread_num_) * 2) {
               std::this_thread::sleep_for(std::chrono::milliseconds(100));
             }
             pool_ptr->enqueue(geneTask(spec_set_ptr_vec[i], selected_prsm_ptrs,
