+++ src/search/varptmsearch/var_ptm_search_processor.cpp
@@ -17,6 +17,7 @@
 
 #include <algorithm>
 #include <cmath>
+#include <cstddef>
 
 #include "common/base/prot_mod_util.hpp"
 #include "common/thread/simple_thread_pool.hpp"
@@ -263,7 +264,7 @@ void VarPtmSearchProcessor::process() {
         if (comp_selected_prsm_ptrs.size() > 0) {
           // LOG_DEBUG("start processing one spectrum.");
           for (size_t k = 0; k < spec_set_vec.size(); k++) {
-            while (pool_ptr->getQueueSize() >= mng_ptr_->thread_num_ * 2) {
+            while (pool_ptr->getQueueSize() >= static_cast<std::size_t>(mng_ptr_->thread_num_) * 2) {
               std::this_thread::sleep_for(std::chrono::milliseconds(100));
             }
             pool_ptr->enqueue(geneTask(
@@ -283,7 +284,7 @@ void VarPtmSearchProcessor::process() {
         if (pref_selected_prsm_ptrs.size() > 0) {
           // LOG_DEBUG("start processing one spectrum.");
           for (size_t k = 0; k < spec_set_vec.size(); k++) {
-            while (pool_ptr->getQueueSize() >= mng_ptr_->thread_num_ * 2) {
+            while (pool_ptr->getQueueSize() >= static_cast<std::size_t>(mng_ptr_->thread_num_) * 2) {
               std::this_thread::sleep_for(std::chrono::milliseconds(100));
             }
             pool_ptr->enqueue(geneTask(
@@ -303,7 +304,7 @@ void VarPtmSearchProcessor::process() {
         if (suff_selected_prsm_ptrs.size() > 0) {
           // LOG_DEBUG("start processing one spectrum.");
           for (size_t k = 0; k < spec_set_vec.size(); k++) {
-            while (pool_ptr->getQueueSize() >= mng_ptr_->thread_num_ * 2) {
+            while (pool_ptr->getQueueSize() >= static_cast<std::size_t>(mng_ptr_->thread_num_) * 2) {
               std::this_thread::sleep_for(std::chrono::milliseconds(100));
             }
             pool_ptr->enqueue(geneTask(
@@ -323,7 +324,7 @@ void VarPtmSearchProcessor::process() {
         if (internal_selected_prsm_ptrs.size() > 0) {
           // LOG_DEBUG("start processing one spectrum.");
           for (size_t k = 0; k < spec_set_vec.size(); k++) {
-            while (pool_ptr->getQueueSize() >= mng_ptr_->thread_num_ * 2) {
+            while (pool_ptr->getQueueSize() >= static_cast<std::size_t>(mng_ptr_->thread_num_) * 2) {
               std::this_thread::sleep_for(std::chrono::milliseconds(100));
             }
             pool_ptr->enqueue(geneTask(spec_set_vec[k],
